# src/main/java/br/com/claudsan/store/adapter/dto/ItemViewDTO.java
package br.com.claudsan.store.adapter.dto;

import lombok.*;

import java.io.Serializable;

@Builder
@Getter
@Setter
public class ItemViewDTO implements Serializable {
    private Long id;
    private String name;
    private Long quantity;

}